+++ src/main/java/lotto/util/Validator.java
@@ -1,0 +1,12 @@
+package lotto.util;
+
+public class Validator {
+
+    static int parseToInt(String input) {
+        try {
+            return Integer.parseInt(input);
+        } catch (NumberFormatException e) {
+            throw new IllegalArgumentException("[ERROR] 숫자만 입력해주세요.");
+        }
+    }
+}
